Fix failing tests and do cleanup
