+++ commercetools.Sdk/Tests/commercetools.Sdk.Serialization.Tests/AttributeDeserializationTests.cs
@@ -55,6 +55,15 @@ namespace commercetools.Sdk.Serialization.Tests
         }
 
         [Fact]
+        public void DeserializeDateAttribute()
+        {
+            var serializerService = this.serializationFixture.SerializerService;
+            string serialized = File.ReadAllText("Resources/Attributes/Date.json");
+            ProductVariant deserialized = serializerService.Deserialize<ProductVariant>(serialized);
+            Assert.IsAssignableFrom<Attribute<DateTime>>(deserialized.Attributes[0]);
+        }
+        
+        [Fact]
         public void DeserializeDateTimeAttribute()
         {
             ISerializerService serializerService = this.serializationFixture.SerializerService;
@@ -283,5 +292,56 @@ namespace commercetools.Sdk.Serialization.Tests
             Assert.True(firstNestedAttr.Value[0].IsTextAttribute());
             Assert.True(firstNestedAttr.Value[3].IsMoneyAttribute());
         }
+        
+        [Fact]
+        public void DeserializeDateAsTextAttribute()
+        {
+            var config = new SerializationConfiguration
+            {
+                DeserializeDateAttributesAsString = true
+            };
+            var serializerService = this.serializationFixture.BuildSerializerServiceWithConfig(config);
+            
+            var serialized = @"
+                {
+                    ""id"": 1,
+                    ""key"": ""newKey"",
+                    ""attributes"": [
+                                        {
+                                            ""name"": ""text-attribute"",
+                                            ""value"": ""2021-10-12""
+                                        }
+                                    ]
+                }
+            ";
+            var deserialized = serializerService.Deserialize<ProductVariant>(serialized);
+            Assert.IsAssignableFrom<Attribute<string>>(deserialized.Attributes[0]);
+        }
+        
+        [Fact]
+        public void DeserializeDateTimeAsTextAttribute()
+        {
+            var config = new SerializationConfiguration
+            {
+                DeserializeDateTimeAttributesAsString = true
+            };
+            var serializerService = this.serializationFixture.BuildSerializerServiceWithConfig(config);
+
+            var serialized = @"
+                {
+                    ""id"": 1,
+                    ""key"": ""newKey"",
+                    ""attributes"": [
+                                        {
+                                            ""name"": ""text-attribute"",
+                                            ""value"": ""2021-10-12 05:50:06""
+                                        }
+                                    ]
+                }
+            ";
+            var deserialized = serializerService.Deserialize<ProductVariant>(serialized);
+            Assert.IsAssignableFrom<Attribute<string>>(deserialized.Attributes[0]);
+        }
+
     }
 }
